Lazy-load vxlapi64 DLL
The use-case is building a program that has Vector support, but
doesn't require the user to have the DLL installed if they're not
using any Vector functionality (e.g. because they only use panda).

The simple solution would've been to ship the DLL with your
software (which I believe to be the standard approach for PCAN), but
that is explicitly forbidden by the license:

> The DLLs (dynamic linked libraries) of the XL Driver Library may not
> be distributed solely or as part of any programs without a previous
> written individual agreement with Vector. For such individual use
> cases, please contact support@vector.com.

Given the above, I also wonder whether the dll should be removed from
this repository, but I will leave that up to the maintainer's
discretion.

I assume that vxlapi64 is thread-safe (their factsheet suggests that),
but they do not explicitly confirm nor deny this in any of their
documentation.

The build script had to be modified to only generate rust code for the
parts of the header file that are actually used, to prevent build
times in the minutes to hours range.

The Error type now no longer implements PartialEq, since the
libloading Error type doesn't, and it's not common for error types in
general.